faculty role api changes complete

+++ Controllers/CourseController.js
@@ -13,29 +13,55 @@ export const courseCreate = {
             !req.body.topics ||
             !req.body.courseduration ||
             !req.body.dailytime ||
-            !req.body.price
+            !req.body.price ||
+            !req.currUser.usertype
         ) {
             return res.status(400).send("Please Fill all the Fields");
         }
         next();
     },
     controller: async (req, res) => {
-        try {
-            const newCohort = await Course.create({
-                title: req.body.title,
-                description: req.body.description,
-                topics: req.body.topics,
-                courseduration: req.body.courseduration,
-                dailytime: req.body.dailytime,
-                price: req.body.price
-            })
-            return res.status(201).send({
-                "message": "Cohort Creation Successful",
-                ...newCohort._doc
-            });
-        } catch (e) {
-            console.log(e);
-            return res.status(500).send("Cohort Creation Failed");
+        if (req.currUser.usertype === "owner") {
+            try {
+                const newCourse = await Course.create({
+                    title: req.body.title,
+                    description: req.body.description,
+                    topics: req.body.topics,
+                    courseduration: req.body.courseduration,
+                    dailytime: req.body.dailytime,
+                    price: req.body.price,
+                    faculty: req.currUser._id,
+                    status: "approved"
+                })
+                return res.status(201).send({
+                    "message": "Cohort Creation Successful",
+                    ...newCourse._doc
+                });
+            } catch (e) {
+                console.log(e);
+                return res.status(500).send("Cohort Creation Failed");
+            }
+        } else {
+            try {
+                const newCourse = await Course.create({
+                    title: req.body.title,
+                    description: req.body.description,
+                    topics: req.body.topics,
+                    courseduration: req.body.courseduration,
+                    dailytime: req.body.dailytime,
+                    price: req.body.price,
+                    faculty: req.currUser._id,
+                    status: "pending"
+                })
+                return res.status(201).send({
+                    "message": "Cohort Creation Successful",
+                    ...newCourse._doc
+                });
+            } catch (e) {
+                console.log(e);
+                return res.status(500).send("Course Creation Failed");
+            }
+
         }
     }
 }
@@ -83,7 +109,8 @@ export const courseUpdate = {
             !req.body.topics ||
             !req.body.courseduration ||
             !req.body.dailytime ||
-            !req.body.price
+            !req.body.price ||
+            !req.currUser.usertype
         ) {
             return res.status(400).send("Please Fill all the Fields");
         } else if (!req.params.id) {
@@ -99,13 +126,19 @@ export const courseUpdate = {
                 return res.status(400).send("Coures is Not found");
             }
 
+            if (findCourse.faculty !== req.currUser._id.toString()) {
+                return res.status(400).send("You can't update the course details")
+            }
+
             const updateCourse = await Course.findByIdAndUpdate(req.params.id, {
                 title: req.body.title,
                 description: req.body.description,
                 topics: req.body.topics,
                 courseduration: req.body.courseduration,
                 dailytime: req.body.dailytime,
-                price: req.body.price
+                price: req.body.price,
+                faculty: req.currUser._id,
+                status: "pending"
             }, { new: true })
             return res.status(200).send({
                 "message": "Course updation Successful",
@@ -118,6 +151,41 @@ export const courseUpdate = {
     }
 }
 
+export const approveCourse = {
+    validator: async (req, res, next) => {
+        if (!req.body._id) {
+            return res.status.send("Invalid Course");
+        }
+        next();
+    },
+    controller: async (req, res) => {
+        console.log(req.currUser);
+        if (req.currUser.usertype === "owner") {
+            try {
+                const findCourse = await Course.findById(req.body._id);
+
+                if (!findCourse) {
+                    return res.status(400).send("Coures is Not found");
+                }
+
+                const updateCourse = await Course.findByIdAndUpdate(req.body._id, {
+                    status: "approved"
+                }, { new: true })
+
+                return res.status(200).send({
+                    "message": "Course approved",
+                    ...updateCourse._doc
+                });
+            } catch (e) {
+                console.log(e);
+                return res.status(500).send("Course approve Failed");
+            }
+        } else {
+            return res.status(400).send("You can't approve the course");
+        }
+    }
+}
+
 export const courseDelete = {
     validator: async (req, res, next) => {
         if (!req.params.id) {
@@ -126,28 +194,32 @@ export const courseDelete = {
         next();
     },
     controller: async (req, res) => {
-        try {
-            const findCourse = await Course.findById(req.params.id);
+        if (req.currUser.usertype === "owner") {
+            try {
+                const findCourse = await Course.findById(req.params.id);
 
-            if (!findCourse) {
-                return res.status(400).send("Cohort is Not found");
-            }
-
-            if (findCourse.batches.length !== 0) {
-                for (let i = 0; i < findCourse.batches.length; i++) {
-                    await Batch.findByIdAndDelete(findCourse.batches[i]);
+                if (!findCourse) {
+                    return res.status(400).send("Cohort is Not found");
                 }
+
+                if (findCourse.batches.length !== 0) {
+                    for (let i = 0; i < findCourse.batches.length; i++) {
+                        await Batch.findByIdAndDelete(findCourse.batches[i]);
+                    }
+                }
+
+                await Course.findByIdAndDelete(req.params.id)
+
+                return res.status(200).send(
+                    "Course Deleted Successfully",
+                );
+
+            } catch (e) {
+                console.log(e);
+                return res.status(500).send("Course Deletion Failed");
             }
-
-            await Course.findByIdAndDelete(req.params.id)
-
-            return res.status(200).send(
-                "Course Deleted Successfully",
-            );
-
-        } catch (e) {
-            console.log(e);
-            return res.status(500).send("Course Deletion Failed");
+        } else {
+            return res.status(400).send("You can't Delete the course")
         }
     }
 }
@@ -155,7 +227,7 @@ export const courseDelete = {
 export const getAllCourses = {
     controller: async (req, res) => {
         try {
-            const findCourses = await Course.find();
+            const findCourses = await Course.find({ status: "approved" });
 
             if (findCourses.length == 0) {
                 return res.status(200).send("Courses Not Found")
@@ -183,6 +255,42 @@ export const getAllCourses = {
     }
 }
 
+
+export const getAllCoursesOfFaculty = {
+    controller: async (req, res) => {
+        try {
+            let findCourses;
+            if (req.currUser.usertype === "owner") {
+                findCourses = await Course.find();
+            } else {
+                findCourses = await Course.find({ faculty: req.currUser._id });
+            }
+
+            if (findCourses.length == 0) {
+                return res.status(200).send("Courses Not Found")
+            }
+
+            // let timeSlots = {};
+            // for (let i = 0; i < findCourses.length; i++) {
+            //     let stamps = findCourses[i].slots;
+            //     // console.log(findCourses[i]);
+            //     timeSlots[findCourses[i]._id] = [];
+            //     for (let j = 0; j < stamps.length; j++) {
+            //         let slotVal = await Slot.findById(stamps[i]);
+            //         timeSlots[findCourses[i]._id].push(slotVal);
+            //     }
+            // }
+
+            // console.log(timeSlots);
+
+            return res.status(201).send(findCourses);
+
+        } catch (e) {
+            console.log(e);
+            return res.status(500).send("Courses Fetching Failed");
+        }
+    }
+}
 
 
 // Batch APIs
@@ -356,12 +464,6 @@ export const batchDelete = {
         }
     }
 }
-
-
-
-
-
-
 
 export const getAllBatches = {
     controller: async (req, res) => {
